AdminService/requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
astroid==2.7.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
Authlib==1.2.1
Babel==2.12.1
backports.zoneinfo==0.2.1
bangla==0.0.2
black==23.3.0
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.1
bokeh==1.4.0
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clean-fid==0.1.35
click==8.1.3
clip-anytorch==2.5.2
cmake==3.26.4
contourpy==1.1.0
coqpit==0.0.17
coverage==7.2.7
cryptography==41.0.2
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.8
decorator==5.1.1
dnspython==2.4.2
docker-pycreds==0.4.0
docopt==0.6.2
ecdsa==0.18.0
einops==0.6.1
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-sso==0.7.2
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
g2pkk==0.1.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.20.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.56.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflect==5.6.0
install==1.3.5
isort==5.12.0
itsdangerous==2.1.2
jamo==0.4.1
jax==0.4.13
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jose==1.0.0
jsonlines==1.2.0
jsonmerge==1.9.0
jsonschema==4.17.3
k-diffusion==0.0.15
keras==2.12.0
kiwisolver==1.4.4
kornia==0.6.12
lazy-object-proxy==1.9.0
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
lit==16.0.6
llvmlite==0.38.1
mailjet==1.4.1
mailjet-rest==1.3.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.6.1
mecab-python3==1.0.5
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
nose2==0.13.0
num2words==0.5.12
numba==0.55.1
numpy==1.21.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.6
packaging==23.1
pandas==2.0.2
passlib==1.7.4
pathspec==0.11.1
pathtools==0.1.2
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pooch==1.6.0
protobuf==4.24.3
psutil==5.9.5
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydub==0.25.1
PyJWT==2.8.0
pylint==2.10.2
pyloudnorm==0.1.1
pynndescent==0.5.10
pyOpenSSL==23.2.0
pyparsing==3.1.0
pypinyin==0.49.0
pyrsistent==0.19.3
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
pyvi==0.1.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.26.0
setproctitle==1.3.2
six==1.16.0
sklearn-crfsuite==0.3.6
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
sqladmin==0.15.1
SQLAlchemy==2.0.20
starlette==0.27.0
stripe==6.7.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
trainer==0.0.20
trampoline==0.1.2
transformers==4.30.2
triton==2.0.0
-e git+https://github.com/coqui-ai/TTS@4cf86523922f3a9640fe919894d1e43743e63d9e#egg=TTS
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
umap-learn==0.5.1
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
wandb==0.15.4
watchfiles==0.20.0
wcwidth==0.2.6
wheel
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.12.1
WTForms==3.0.1
python-json-logger==2.0.7
yarl==1.9.2
zipp==3.15.0
